Finish this

+++ spec/support/helpers/results_helper.rb
@@ -8,9 +8,10 @@ TOKEN_FILE_PATH = File.join(File.dirname(__FILE__), "../../../token.txt")
 module Helpers
   # helpers for post results
   class ResultsHelper
-    def sign_in(user, password)
+    def sign_in
       response = RestClient.post("#{ENV.fetch('RESULTS_URL', nil)}/sign-in.json",
-                                 user_login: { email: user, password: })
+                                 user_login: { email: ENV.fetch("RESULTS_USER", nil),
+                                 password: ENV.fetch("RESULTS_PASSWORD", nil) })
       @auth_token = JSON.parse(response.body)["auth_token"]
 
       file = File.open(TOKEN_FILE_PATH, "w")
@@ -34,12 +35,14 @@ module Helpers
     end
 
     def sign_out
+      return !File.exist?(TOKEN_FILE_PATH)
+
       RestClient.delete("#{ENV.fetch('RESULTS_URL', nil)}/sign-out.json",
                         "Authorization" => "Token token=#{File.read(TOKEN_FILE_PATH)}")
     rescue Errno::ECONNREFUSED, RestClient::Unauthorized
       # don't care if can't connect or if we're unauthed
     ensure
-      File.delete(TOKEN_FILE_PATH)
+      File.delete(TOKEN_FILE_PATH) if File.exist?(TOKEN_FILE_PATH)
     end
 
     private
